requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
ipython>=8.0.0  # Optional: for display_summary functionality
flask>=2.3.0 